requirements.txt
langgraph==0.4.8
langchain==0.3.25
langchain-openai==0.3.21
langchain-postgres==0.0.14
psycopg2-binary==2.9.10
python-dotenv==1.1.0 